backend/requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic-settings>=2.2.1
boto3>=1.34.84
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
openai>=1.16.2
scikit-learn>=1.4.1.post1
razorpay>=1.4.1
python-dotenv>=1.0.1
